+++ internal/cmd/sentinel.go
@@ -179,6 +179,7 @@ func runSentinel(cmd *cobra.Command, args []string) error {
 			tunnelServer.OnConnect = manager.OnTunnelConnect
 			tunnelServer.OnDisconnect = manager.OnTunnelDisconnect
 			manager.SetTunnelRegistry(registry)
+			manager.SetTunnelPolicy(tunnelPolicy)
 			manager.SetHTTPSListener(connMux.HTTPSChanListener())
 
 			sigChan := make(chan os.Signal, 1)
@@ -263,6 +264,7 @@ func runSentinel(cmd *cobra.Command, args []string) error {
 		tunnelServer.OnConnect = manager.OnTunnelConnect
 		tunnelServer.OnDisconnect = manager.OnTunnelDisconnect
 		manager.SetTunnelRegistry(registry)
+		manager.SetTunnelPolicy(tunnelPolicy)
 		manager.SetHTTPSListener(connMux.HTTPSChanListener())
 
 		// Graceful shutdown on signals